LayoutTests/svg/animations/animated-svg-image-removed-from-document-paused-expected.txt

Tests that animated SVG images are paused when removed from the document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.isImageAnimating(imageA) is true
PASS internals.isImageAnimating(imageB) is true
imageA.remove()
PASS internals.isImageAnimating(imageB) is true
PASS internals.isImageAnimating(imageB) is true
imageB.remove()
PASS internals.isImageAnimating(imageA) is false
PASS internals.isImageAnimating(imageB) is false
document.body.appendChild(imageA)
PASS internals.isImageAnimating(imageA) is true
document.body.appendChild(imageB)
PASS internals.isImageAnimating(imageB) is true
PASS successfullyParsed is true

TEST COMPLETE
